#5215: Split `Match Check` into two inspections
`Match Check` is now divided into two inspections:
* `RsNonExhaustiveMatchInspection` (E0004 error)
* `RsUselessMatchArmInspection` (unreachable_patterns warning)
