--- src/main/resources/inspectionDescriptions/RsMatchCheck.html
@@ -1,5 +1,0 @@
-<html>
-<body>
-Checks that <code>match</code> expression is exhaustive and doesn't have useless arms.
-</body>
-</html>
