+++ src/main/resources/inspectionDescriptions/RsNonExhaustiveMatch.html
@@ -1,0 +1,5 @@
+<html>
+<body>
+Checks that <code>match</code> expression is exhaustive, i.e. all possible patterns are covered.
+</body>
+</html>
