+++ src/main/resources/inspectionDescriptions/RsUnreachablePatterns.html
@@ -1,0 +1,5 @@
+<html>
+<body>
+Checks that <code>match</code> expression doesn't have unreachable patterns.
+</body>
+</html>
